refactor: move OperationMode to app/models, consolidate button creation into ButtonFactory
- Extract OperationMode enum from base_mods_panel.py into
  app/models/operation_mode.py to eliminate circular imports
- Add _create_dropdown_button() helper to remove duplicate
  QPushButton + QMenu + exec pattern across 3 methods
- Move create_steam_button() and create_refresh_button() logic
  fully into the factory (previously delegated to panel)
- Refactor create_steamcmd_button() and create_select_all_button()
  to use the new dropdown helper
- Remove now-unused panel methods: _create_button,
  _create_standardized_button, _create_refresh_button,
  _create_steam_button
