+++ app/models/operation_mode.py
@@ -1,0 +1,10 @@
+from __future__ import annotations
+
+from enum import Enum
+
+
+class OperationMode(Enum):
+    """Enumeration for mod update operation modes."""
+
+    STEAMCMD = "SteamCMD"
+    STEAM = "Steam"
